requirements.txt
PyYAML>=6.0
pandas>=2.0.0
numpy>=1.24.0
